pyproject.toml
[tool.poetry]
name = "rosatom"
version = "0.1.0"
description = ""
authors = ["SemyonSinchenko <semyonsinchenko@yahoo.com>"]

[[tool.poetry.source]]
name = "pypi_"
url = "https://pypi.org/simple/"

[[tool.poetry.source]]
name = "qubo"
url = "https://repo.qboard.tech/"

[tool.poetry.dependencies]
python = ">=3.7.1,<3.10"
numpy = "^1.21.3"
scipy = "^1.7.1"
matplotlib = "^3.4.3"
ipython = "^7.28.0"
black = "^21.9b0"
pandas = "^1.3.4"
networkx = "^2.6.3"
pdoc = "^8.0.0"
# qboard-client # -- doen't work because "Invalid hash for qboard-client (0.1.0) using archive qboard-client-0.1.0.tar.gz"
# shoul be installed directly using pip install
"stomp.py" = "^7.0.0"
pydantic = "^1.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
